Append hint to Progress Bar instead of body
Second pass at attempting to support allowing the user to control the
color of the progress bar hover. Currently it's hardcoded to be yellow.
Some users, like myself, might want to match the hover color with the
progress bar color.

My first attempt I added the customClass to the hint div. However,
usmonster suggested to better support backwards compatibility that I
instead try to append the hint div to the bar div since the bar div
already has the customClass. This would give developers control over the
hint div with a selector like .bar.ganttRed .fn-gantt-hint. These
changes accomplish these use cases.

However, I think the changes were more involved than we originally
anticipated. There were two main issues I had to work around: position
and z-index. First, now that we append to the bar div and not body, the
left/top positioning had to be updated since the hint div is now
relative to the bar div. So I had to use event.offsetX/Y, but those
values were only available in Chrome and IE and not in Firefox.
According to this jquery bug, http://bugs.jquery.com/ticket/8523, it's
best to calculate the offset when offsetX/Y are not present. The second
issue, was a z-index issue. I was curious to see how the hover worked if
I had another progress bar right below another. Turns out that when you
do this, with these changes, the hover is behind the other progress bar
even though the z-index is correct. It appears that when the hint div
was moved to the bar div this caused the issue. Moving it back to being
appended to the body I don't see this issue. The solution was to remove
the z-index for the .fn-gantt .bar style. This didn't seem to have any
negative side effects, but it may not be the solution we want.

I also applied an additional offset of 15 for both top and left.
Previously we were only adding 15 for top and that was only on the
mousemove. I added 15 to both top and left, and during mouseover and
mousemove. Also doing it in mouseover prevents the hint from jumping,
which I only noticed in IE.

Finally, at the suggestion of usmonster, I replaced hint.remove() with
hint.detach() in mouseout.

I've tested all these changes in Chrome 30, Firefox 24, and IE9.